Feat : Complete Pilot Run
BODY : Parameters such as learning rate and the weight adjustment for
over represented 'Fish' class were optimized in this run.  It appears
that a beta_rate of about 0.13 (the proportion of NoF to total images)
is appropriate for accuracy to straddle 50%.  Learning rates of 0.001
and 0.005 were not sufficient to drive efficient learning.  Only the
bias parameters seemed to change.  Learning was evident at a learning
rate of 0.1.

The next step is to try a multi-task approach.

ref : FishNoF