Feat : Implement FishFinder model
BODY : FishFinder is a NN that receives coarse images of high-resolution
jpegs and learns whether a fish is present within the image, along with
the top left coordinates for a bounding box and the scale required to
fit the entire fish into a standard sized fovea.  The yx coordinates
are given as a proportion of the coarse image dimensions (i.e. 0 to 1).

These tasks are trained together, with the final output from dense
connected layers being subjected to both a FishNoF classifier and a
box classifier that returns [scale, y_offset, x_offset].

Several tricks were required to implement this multitasking learning:
     1. Box labels that are missing get imputed with a dummy variable
     ('-1') for both scale and coordinates.
     2.  Along with the box labels, a weights vector is also supplied,
     which is populated with zeros or ones, based on whether an actual
     label exists for the example.  Thus any cross entropy derived from
     the dummy variables gets negated, and will not be utilized for
     gradient calculation and training.  For 'NoF' examples, the labels
     will always be missing.

Annotations for the box specs are added during the run in the
`Box_annotation.ipynb` notebook file, using the `boxit` fn.  However
because the validation set was generated at the beginning of the run,
the valid cross entropy for box prediction remains at zero as none of
the valid samples had labels at the onset of training.

The beta parameter is used to aid in training of the FishNoF
classification.  Because the dataset is 87% fish and 13% NoF, the
imbalance causes the network to bias prediction towards the 'Fish'
classification.  By providing a coefficent less than one for weighting
of the cross entropy calcualtion, the fish examples become less
important in the gradient calculation, whereas the NoF examples become
more important.  This parameter is not completely necessary, as
eventually the network will learn the features the represent NoF and
the distribution of FishNoF logit probabilities will diverge away
from high confidence Fish predictions.