Feat : Train v1.1
BODY :

I've lost my way a little bit tweaking parameters and features of this
training session.  In general, the architecture is good and here are
some items I have learned from this pilot run :

     * High learning rates compress the FishNoF logits.  Try 1e-3 to
     start the next implementation.
     * Regularization on the classifying layers is not necessary.  Keep
     something very low at first for the dense layers.  (1e-8).
     * The bulk of training time was simply loading the image with the
     CPU.  Keep this in mind for future architectures.

I need to add a 'prediction' chunk where the network is used to populate
a dictionary where each image is given the current predicted value
for FishNoF as well as box specs.  This can be taken in a different
session for FishyFish model training.

ref : v1.1