Feat : Train FishyFish
BODY :

100 epochs of training was performed.  Then a quick and dirty
session call was run to make the predictions from test images.

I had to append the stg1 benchmark values to the stg2 test images
as I was out of time to process those images as well.  I did not
realize they were necessary for submission.  In any case,
as long as the actual test images are from stage 2 my models predictions
will be used.

ref v1.0